Update form.php
Hi, fix for required email not being required in for (displays blue in stead of red. Check for required is currently coded but is only checked when form is posted. With this change the field is checked BEFORE posting :)

Second fix is the addition of the warning when not logged in. This is implemented on the kunena component and WAS implemented in KunenaDiscuss 3.x. This reintroduces this feature :)

Hope this helps!
Keep up the good work, you are doing an awesome job!